+++ SamplingApp/Sampler/UniformSampler.cpp
@@ -9,7 +9,7 @@ namespace {
  * Generates uniformly distributed Samples in the range [0,1].
  *
  */
-void UniformSampler::generateSamples(int nSamplesSqrt, std::vector<Math::Vec2f>& sampleStorage) {
+void UniformSampler::generateSamples(size_t nSamplesSqrt, std::vector<Math::Vec2f>& sampleStorage) {
 	float spacing = 1.f/(2*nSamplesSqrt);
 	for (size_t i = 0; i < nSamplesSqrt; i++) {
 		for (size_t j = 0; j < nSamplesSqrt; j++) {
